Refactored the utils part as an NPM package; Moved index.ts to cli/; Fixed package.json to use CLI from another directory
