+++ src/utils/buildGrid.ts
@@ -1,6 +1,7 @@
 import pointGrid from '@turf/point-grid';
-import { Feature, Geometry, Position } from '@turf/helpers/dist/js/lib/geojson';
+import { Feature, Geometry } from '@turf/helpers/dist/js/lib/geojson';
+import { Position, RectangleType } from './declarations';
 
 export const buildGrid = (
-  extent: [number, number, number, number], cellSide: number
-): Position[] => pointGrid(extent, cellSide).features.map((feature: Feature<Geometry>) => feature.geometry.coordinates);
+  extent: RectangleType, cellSize: number
+): Position[] => pointGrid(extent, cellSize).features.map((feature: Feature<Geometry>) => feature.geometry.coordinates);
