+++ src/utils/declarations.ts
@@ -1,0 +1,8 @@
+export interface Cell {
+  cell: number;
+}
+
+export type Position = number[];
+export type PointPositionType = [number, number]
+export type RectangleType = [number, number, number, number];
+
