+++ src/utils/index.ts
@@ -2,5 +2,6 @@ export * from './buildExtent';
 export * from './buildGpx';
 export * from './buildGrid';
 export * from './buildNamesDict';
+export * from './declarations';
 export * from './parsePositionFromString';
 export * from './prepareData';
