+++ src/utils/prepareData.ts
@@ -1,20 +1,19 @@
-import { Position } from '@turf/helpers/dist/js/lib/geojson';
 import distance from '@turf/distance';
 import { parsePositionFromString } from './parsePositionFromString';
 import { buildExtent } from './buildExtent';
 import destination from "@turf/destination";
+import { Cell, PointPositionType, Position, RectangleType } from './declarations';
 
-export interface RowData {
+export interface RowData extends Cell {
   center?: string;
   radius?: number;
-  cell?: number;
   fromPoint?: string;
   toPoint?: string;
 }
 
-export function getExtent(data: RowData): [number,number,number,number] {
+export function getExtent(data: RowData): RectangleType {
   if (data.center && data.radius) {
-    const centerPosition: Position = parsePositionFromString(data.center);
+    const centerPosition: PointPositionType = parsePositionFromString(data.center);
 
     return buildExtent(centerPosition, data.radius);
   } else if (data.fromPoint && data.toPoint && data.cell) {
@@ -28,7 +27,7 @@ export function getExtent(data: RowData): [number,number,number,number] {
 
 export function getDefaultFileName(data: RowData): string {
   if (data.center && data.radius && data.cell) {
-    const centerPosition: Position = parsePositionFromString(data.center);
+    const centerPosition: PointPositionType = parsePositionFromString(data.center);
 
     return `${centerPosition[1]}_${centerPosition[0]}_${data.radius}x${data.radius}_${data.cell}`;
   } else if (data.fromPoint && data.toPoint && data.cell) {
@@ -40,10 +39,10 @@ export function getDefaultFileName(data: RowData): string {
   }
 }
 
-function minMaxPoint(firstPoint: string, secondPoint: string, cellSize: number): {minPoint: [number, number], maxPoint: [number, number]} {
+function minMaxPoint(firstPoint: string, secondPoint: string, cellSize: number): {minPoint: PointPositionType, maxPoint: PointPositionType} {
   const firstPosition: Position = parsePositionFromString(firstPoint);
   const secondPosition: Position = parsePositionFromString(secondPoint);
-  const minPoint: [number, number] = [
+  const minPoint: PointPositionType = [
     Math.min(firstPosition[0], secondPosition[0]),
     Math.min(firstPosition[1], secondPosition[1]),
   ];
